pyproject.toml
[project]
name = "akari"
version = "0.1.0"
description = "A feature-rich Discord bot with AI capabilities"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Helian Nuits", QQ = "903928770"}
]
dependencies = [
    "beautifulsoup4>=4.12.2",
    "discord-py>=2.5.2",
    "google-generativeai>=0.8.5",
    "lxml>=4.9.3",
    "meme-generator<0.2.0",
    "psutil>=7.0.0",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.2.1",
    "python-dotenv>=1.1.0",
    "aiohttp>=3.9.3",
    "typing-extensions>=4.10.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.2.0",
    "isort>=5.13.2",
    "mypy>=1.9.0",
    "ruff>=0.3.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
    "pytest-cov>=4.1.0",
]

[project.scripts]
akari = "akari.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["akari"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
plugins = ["pydantic.mypy"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "B", "I"]
ignore = []

[tool.ruff.isort]
known-first-party = ["akari"]
